Remove python circular dependency (#103)
