+++ sdk-python/littlehorse/__init__.py
@@ -1,10 +1,10 @@
-from littlehorse.proto_utils import to_variable_value as _to_variable_value
-from littlehorse.user_utils import create_workflow_spec as _create_workflow_spec
-from littlehorse.user_utils import (
+from littlehorse.utils import to_variable_value as _to_variable_value
+from littlehorse.workflow import create_workflow_spec as _create_workflow_spec
+from littlehorse.workflow import (
     create_external_event_def as _create_external_event_def,
 )
-from littlehorse.user_utils import create_task_def as _create_task_def
-from littlehorse.user_utils import start as _start
+from littlehorse.workflow import create_task_def as _create_task_def
+from littlehorse.worker import start as _start
 
 
 start = _start
