--- sdk-python/littlehorse/common_utils.py
@@ -1,32 +1,0 @@
-"""Collections of utils for general proposes"""
-
-from pathlib import Path
-import sys
-from typing import Union
-
-
-def read_binary(file_path: Union[str, Path]) -> bytes:
-    """Read a file to bytes.
-
-    Args:
-        file_path (Union[str, Path]): File location.
-
-    Returns:
-        bytes: File bytes.
-    """
-    with open(file_path, "rb") as file_input:
-        return file_input.read()
-
-
-def get_event_loop_is_deprecated() -> bool:
-    """Verify if asyncio.get_event_loop()
-    is deprecated for the current python version.
-
-    https://docs.python.org/3/library/asyncio-eventloop.html#asyncio.get_event_loop.
-
-    Use asyncio.run(main()).
-
-    Returns:
-        bool: True if python is greater or equal to 3.10.
-    """
-    return sys.version_info >= (3, 10, 0)
