+++ sdk-python/littlehorse/worker.py
@@ -1,7 +1,9 @@
 import asyncio
 from datetime import datetime
+import functools
 from inspect import Parameter, signature, iscoroutinefunction
 import logging
+import signal
 from typing import Any, AsyncIterator, Callable
 from littlehorse.config import LHConfig
 from littlehorse.exceptions import (
@@ -16,13 +18,10 @@ from littlehorse.model.service_pb2 import (
     ReportTaskRun,
     ScheduledTask,
 )
+from google.protobuf.timestamp_pb2 import Timestamp
 from littlehorse.model.task_def_pb2 import TaskDef
-from littlehorse.proto_utils import (
-    to_variable_value,
-    to_type,
-    extract_value,
-    timestamp_now,
-)
+from littlehorse.utils import extract_value, to_variable_value
+from littlehorse.utils import to_type
 
 REPORT_TASK_DEFAULT_RETRIES = 5
 HEARTBEAT_DEFAULT_INTERVAL = 5
@@ -283,9 +282,12 @@ class LHConnection:
 
         self._schedule_task_semaphore.release()
 
+        current_time = Timestamp()
+        current_time.GetCurrentTime()
+
         task_result = ReportTaskRun(
             task_run_id=task.task_run_id,
-            time=timestamp_now(),
+            time=current_time,
             attempt_number=task.attempt_number,
             status=status,
             output=output,
@@ -488,3 +490,23 @@ class LHTaskWorker:
 
         for connection in self._connections.values():
             connection.stop()
+
+
+def shutdown_hook(*workers: LHTaskWorker) -> None:
+    """Add a shutdown hook for multiples workers"""
+
+    def stop_workers(*workers: LHTaskWorker) -> None:
+        for worker in workers:
+            worker.stop()
+
+    loop = asyncio.get_running_loop()
+
+    for sig in (signal.SIGHUP, signal.SIGTERM, signal.SIGINT):
+        loop.add_signal_handler(sig, functools.partial(stop_workers, *workers))
+
+
+async def start(*workers: LHTaskWorker) -> None:
+    """Starts a list of workers"""
+    shutdown_hook(*workers)
+    tasks = [asyncio.create_task(worker.start()) for worker in workers]
+    await asyncio.gather(*tasks)
